Migrate utilDescribe out to its own file
